// src/app/api/models/ActivitiesStudent.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BookingResponseDTO } from './BookingResponseDTO';
import type { OrderResponseForTeacherDTO } from './OrderResponseForTeacherDTO';
export type ActivitiesStudent = {
    ordersHistory?: Array<OrderResponseForTeacherDTO> | null;
    bookingsOftheWeek?: Array<BookingResponseDTO> | null;
};